+++ apps/bridge/hooks/use-transaction-args/cctp-args/use-cctp-bridge-args.ts
@@ -6,8 +6,9 @@ import { useFromChain, useToChain } from "@/hooks/use-chain";
 import { useGraffiti } from "@/hooks/use-graffiti";
 import { useWeiAmount } from "@/hooks/use-wei-amount";
 import { useConfigState } from "@/state/config";
+import { isCctp } from "@/utils/is-cctp";
 
-import { addressToBytes32, isCctpBridgeOperation } from "./common";
+import { addressToBytes32 } from "./common";
 
 export const useCctpArgs = () => {
   const stateToken = useConfigState.useToken();
@@ -30,7 +31,7 @@ export const useCctpArgs = () => {
     !fromCctp ||
     !toCctp ||
     !recipientAddress ||
-    !isCctpBridgeOperation(stateToken)
+    !isCctp(stateToken)
   ) {
     return;
   }
