+++ apps/bridge/hooks/use-transaction-args/deposit-args/use-deposit-args.ts
@@ -1,17 +1,18 @@
 import { useConfigState } from "@/state/config";
+import { isCctp } from "@/utils/is-cctp";
 
 import { useArbitrumDepositArgs } from "./use-arbitrum-deposit-args";
 import { useOptimismDepositArgs } from "./use-optimism-deposit-args";
-import { isCctpBridgeOperation } from "../cctp-args/common";
 
 export const useDepositArgs = () => {
   const withdrawing = useConfigState.useWithdrawing();
   const stateToken = useConfigState.useToken();
+  const fast = useConfigState.useFast();
 
   const optimism = useOptimismDepositArgs();
   const arbitrum = useArbitrumDepositArgs();
 
-  if (withdrawing || !stateToken || isCctpBridgeOperation(stateToken)) {
+  if (withdrawing || !stateToken || isCctp(stateToken) || fast) {
     return;
   }
