+++ apps/bridge/hooks/use-transaction-args/index.ts
@@ -1,10 +1,10 @@
 import { useConfigState } from "@/state/config";
+import { isCctp } from "@/utils/is-cctp";
 
 import { useTokenBalance } from "../use-balances";
 import { useDeployment } from "../use-deployment";
 import { useSelectedToken } from "../use-selected-token";
 import { useWeiAmount } from "../use-wei-amount";
-import { isCctpBridgeOperation } from "./cctp-args/common";
 import { useCctpArgs } from "./cctp-args/use-cctp-bridge-args";
 import { useDepositArgs } from "./deposit-args/use-deposit-args";
 import { forceWithdrawalArgs } from "./force-withdrawal-args";
@@ -33,7 +33,7 @@ export const useTransactionArgs = () => {
     return across;
   }
 
-  if (isCctpBridgeOperation(stateToken)) {
+  if (isCctp(stateToken)) {
     if (forceViaL1) return forceWithdrawalArgs(cctp, deployment);
     return cctp;
   }
