+++ apps/bridge/hooks/use-transaction-args/withdraw-args/use-withdraw-args.ts
@@ -1,8 +1,8 @@
 import { useConfigState } from "@/state/config";
+import { isCctp } from "@/utils/is-cctp";
 
 import { useArbitrumWithdrawArgs } from "./use-arbitrum-withdraw-args";
 import { useOptimismWithdrawArgs } from "./use-optimism-withdraw-args";
-import { isCctpBridgeOperation } from "../cctp-args/common";
 
 export const useWithdrawArgs = () => {
   const withdrawing = useConfigState.useWithdrawing();
@@ -11,7 +11,7 @@ export const useWithdrawArgs = () => {
   const optimism = useOptimismWithdrawArgs();
   const arbitrum = useArbitrumWithdrawArgs();
 
-  if (!withdrawing || !stateToken || isCctpBridgeOperation(stateToken)) {
+  if (!withdrawing || !stateToken || isCctp(stateToken)) {
     return;
   }
